% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_charts.R
\name{available_charts}
\alias{available_charts}
\title{List of available charts in nvd3R}
\usage{
available_charts()
}
\description{
List of available charts in nvd3R
}
\examples{
library(nvd3R)
available_charts()

}
