% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd3Plot.R
\name{nvd3Plot}
\alias{nvd3Plot}
\title{Create NVD3 charts}
\usage{
nvd3Plot(
  data,
  x,
  y,
  col = NULL,
  type = "boxplot",
  xlab = "X-axis",
  ylab = "Y-axis"
)
}
\arguments{
\item{data}{data.frame}

\item{x, y}{variables in quote e.i "var1"}

\item{col}{color using variable}

\item{type}{Chart type (see details)}

\item{xlab, ylab}{Axis label}
}
\description{
Main function to make new NVD3 chart
}
\examples{
library(nvd3R)

### BOXPLOT
mtcars \%>\%
nvd3Plot(
x = "cyl",
y = "mpg",
type = "boxplot",
xlab = "Test1",
ylab = "Test2"
)


### Multibarplot
newdata02 <- data.frame(
var1 = c("A", "B", "C", "D", "A", "B", "C", "D"),
var2 = c(10,15,20,25,30,35,40,98),
label = c("A", "A", "A", "A", "B", "B", "B", "B")
)

newdata02 \%>\%
nvd3Plot(
x = "var1",
y = "var2",
col = "label",
type = "multibarplot"
)


}
